Include MCEq to calculate muon flux
